+++ pkg/ffuf/job.go
@@ -39,7 +39,9 @@ type Job struct {
 	skipQueue            bool
 	currentDepth         int
 	calibMutex           sync.Mutex
+	blacklistMutex       sync.Mutex
 	pauseWg              sync.WaitGroup
+	blacklistChecked     bool
 }
 
 type QueueJob struct {
@@ -429,6 +431,11 @@ func (j *Job) runTask(input map[string][]byte, position int, retried bool) {
 		}
 	}
 	j.pauseWg.Wait()
+
+	// use a list of commonly blacklisted files to detect & filter blacklist responses
+	if j.Config.BlackListDetection {
+		j.DetectBlacklist(HostURLFromRequest(req), input)
+	}
 
 	// Handle autocalibration, must be done after the actual request to ensure sane value in req.Host
 	_ = j.CalibrateIfNeeded(HostURLFromRequest(req), input)